+++ azure-dataplane-tests/src/main/java/com/azure/containers/containerregistry/implementation/models/ArtifactTagPropertiesInternal.java
@@ -5,6 +5,7 @@
 package com.azure.containers.containerregistry.implementation.models;
 
 import com.azure.core.annotation.Fluent;
+import com.azure.core.util.CoreUtils;
 import com.azure.json.JsonReader;
 import com.azure.json.JsonSerializable;
 import com.azure.json.JsonToken;
@@ -343,11 +344,11 @@ public class ArtifactTagPropertiesInternal implements JsonSerializable<ArtifactT
                         } else if ("digest".equals(fieldName)) {
                             deserializedArtifactTagPropertiesInternal.digest = reader.getString();
                         } else if ("createdTime".equals(fieldName)) {
-                            deserializedArtifactTagPropertiesInternal.createdOn
-                                = reader.getNullable(nonNullReader -> OffsetDateTime.parse(nonNullReader.getString()));
+                            deserializedArtifactTagPropertiesInternal.createdOn = reader.getNullable(
+                                nonNullReader -> CoreUtils.parseBestOffsetDateTime(nonNullReader.getString()));
                         } else if ("lastUpdateTime".equals(fieldName)) {
-                            deserializedArtifactTagPropertiesInternal.lastUpdatedOn
-                                = reader.getNullable(nonNullReader -> OffsetDateTime.parse(nonNullReader.getString()));
+                            deserializedArtifactTagPropertiesInternal.lastUpdatedOn = reader.getNullable(
+                                nonNullReader -> CoreUtils.parseBestOffsetDateTime(nonNullReader.getString()));
                         } else if ("changeableAttributes".equals(fieldName)
                             && reader.currentToken() == JsonToken.START_OBJECT) {
                             while (reader.nextToken() != JsonToken.END_OBJECT) {