+++ typespec-extension/src/code-model-builder.ts
@@ -876,34 +876,30 @@ export class CodeModelBuilder {
       let pollingSchema = undefined;
       let finalSchema = undefined;
 
+      let pollingStrategy: Metadata | undefined = undefined;
+      let finalResultPropertySerializedName: string | undefined = undefined;
+
       const verb = httpOperation.verb;
       const useNewPollStrategy = isLroNewPollingStrategy(httpOperation, lroMetadata);
-
-      let pollingStrategy: Metadata | undefined = undefined;
       if (useNewPollStrategy) {
-        // use new experimental OperationLocationPollingStrategy
+        // use OperationLocationPollingStrategy
         pollingStrategy = new Metadata({
           language: {
             java: {
               name: "OperationLocationPollingStrategy",
-              namespace: "com.azure.core.experimental.util.polling",
+              namespace: getJavaNamespace(this.namespace) + ".implementation",
             },
           },
         });
       }
 
       // pollingSchema
-      if (useNewPollStrategy) {
-        // com.azure.core.experimental.models.PollResult
+      if (modelIs(lroMetadata.pollingInfo.responseModel, "OperationStatus", "Azure.Core.Foundations")) {
         pollingSchema = this.pollResultSchema;
       } else {
-        if (modelIs(lroMetadata.pollingInfo.responseModel, "OperationStatus", "Azure.Core.Foundations")) {
-          pollingSchema = this.pollResultSchema;
-        } else {
-          const pollType = this.findResponseBody(lroMetadata.pollingInfo.responseModel);
-          const sdkType = getClientType(this.sdkContext, pollType);
-          pollingSchema = this.processSchemaFromSdkType(sdkType, "pollResult");
-        }
+        const pollType = this.findResponseBody(lroMetadata.pollingInfo.responseModel);
+        const sdkType = getClientType(this.sdkContext, pollType);
+        pollingSchema = this.processSchemaFromSdkType(sdkType, "pollResult");
       }
 
       // finalSchema
@@ -918,6 +914,16 @@ export class CodeModelBuilder {
         const finalType = this.findResponseBody(finalResult);
         const sdkType = getClientType(this.sdkContext, finalType);
         finalSchema = this.processSchemaFromSdkType(sdkType, "finalResult");
+
+        if (
+          useNewPollStrategy &&
+          lroMetadata.finalStep &&
+          lroMetadata.finalStep.kind === "pollingSuccessProperty" &&
+          lroMetadata.finalStep.target
+        ) {
+          // final result is the value in lroMetadata.finalStep.target
+          finalResultPropertySerializedName = this.getSerializedName(lroMetadata.finalStep.target);
+        }
       }
 
       // track usage
@@ -938,7 +944,13 @@ export class CodeModelBuilder {
         }
       }
 
-      op.lroMetadata = new LongRunningMetadata(true, pollingSchema, finalSchema, pollingStrategy);
+      op.lroMetadata = new LongRunningMetadata(
+        true,
+        pollingSchema,
+        finalSchema,
+        pollingStrategy,
+        finalResultPropertySerializedName,
+      );
       return op.lroMetadata;
     }
 
@@ -2091,6 +2103,11 @@ export class CodeModelBuilder {
       extensions = extensions ?? {};
       extensions["x-ms-client-flatten"] = true;
     }
+    const mutability = this.getMutability(prop);
+    if (mutability) {
+      extensions = extensions ?? {};
+      extensions["x-ms-mutability"] = mutability;
+    }
 
     if (prop.kind === "property" && prop.isMultipartFileInput) {
       schema = this.processMultipartFormDataFilePropertySchemaFromSdkType(prop, this.namespace);
@@ -2302,6 +2319,29 @@ export class CodeModelBuilder {
       return !target.visibility.includes(Visibility.Read);
     } else {
       return false;
+    }
+  }
+
+  private getMutability(target: SdkModelPropertyType): string[] | undefined {
+    if (target.kind === "property" && target.visibility) {
+      const mutability: string[] = [];
+      if (target.visibility.includes(Visibility.Create)) {
+        mutability.push("create");
+      }
+      if (target.visibility.includes(Visibility.Update)) {
+        mutability.push("update");
+      }
+      if (target.visibility.includes(Visibility.Read)) {
+        mutability.push("read");
+      }
+      if (mutability.length === 3) {
+        // if all 3 (supported) mutability values are present, there is no need to set the x-ms-mutability
+        return undefined;
+      } else {
+        return mutability;
+      }
+    } else {
+      return undefined;
     }
   }