+++ typespec-tests/src/main/java/com/cadl/longrunning/models/JobResult.java
@@ -7,6 +7,7 @@ package com.cadl.longrunning.models;
 import com.azure.core.annotation.Generated;
 import com.azure.core.annotation.Immutable;
 import com.azure.core.models.ResponseError;
+import com.azure.core.util.CoreUtils;
 import com.azure.json.JsonReader;
 import com.azure.json.JsonSerializable;
 import com.azure.json.JsonToken;
@@ -172,14 +173,14 @@ public final class JobResult implements JsonSerializable<JobResult> {
                 } else if ("status".equals(fieldName)) {
                     deserializedJobResult.status = JobStatus.fromString(reader.getString());
                 } else if ("createdDateTime".equals(fieldName)) {
-                    deserializedJobResult.createdDateTime
-                        = reader.getNullable(nonNullReader -> OffsetDateTime.parse(nonNullReader.getString()));
+                    deserializedJobResult.createdDateTime = reader
+                        .getNullable(nonNullReader -> CoreUtils.parseBestOffsetDateTime(nonNullReader.getString()));
                 } else if ("expirationDateTime".equals(fieldName)) {
-                    deserializedJobResult.expirationDateTime
-                        = reader.getNullable(nonNullReader -> OffsetDateTime.parse(nonNullReader.getString()));
+                    deserializedJobResult.expirationDateTime = reader
+                        .getNullable(nonNullReader -> CoreUtils.parseBestOffsetDateTime(nonNullReader.getString()));
                 } else if ("lastUpdateDateTime".equals(fieldName)) {
-                    deserializedJobResult.lastUpdateDateTime
-                        = reader.getNullable(nonNullReader -> OffsetDateTime.parse(nonNullReader.getString()));
+                    deserializedJobResult.lastUpdateDateTime = reader
+                        .getNullable(nonNullReader -> CoreUtils.parseBestOffsetDateTime(nonNullReader.getString()));
                 } else if ("error".equals(fieldName)) {
                     deserializedJobResult.error = ResponseError.fromJson(reader);
                 } else if ("result".equals(fieldName)) {